Add AST
Lithium increasingly needs to traverse the input code to do some
transformations, and it’s getting unwieldy. Sub-expressions are
accessed with first, second, rest, etc., depending on the type
of the expression, and the code is peppered with these calls,
making functions like collect-closures very hard to grok and
maintain.

Hence, a high-level intermediate representation. For now, we don’t
touch any of the existing compiler code, but the idea is to rewrite
the code to use it. Hopefully it will result in cleaner, simpler
code.