Add a new implementation of collect-closures that uses AST
It should be much simpler to follow than the old one in
lithium.compiler. It’s also divided into two steps
(free variable analysis and the actual collecting) and so
follows Ghuloum’s paper more closely.

The only thing I don’t particularly like is that it
uses an atom under the hood, whereas the old one was
purely functional. But I think it’s a fair price to pay
for simplicity.