Make Lithium-compiled programs debuggable, redo calling convention
Ghuloum advocates using space below SP for temporary storage, which
confuses DOS debuggers (see discussion in [1]). What we should do
instead is `sub sp, [amount of space we need]` at the prolog of each
function, including the main code, so that these temporary locations
are on the stack, rather than below it.

For now, we add a fixed 16 bytes for each stack frame, but eventually
we should allow for however many temporary locations the function
needs (it is inferrable statically).

To reclaim that temporary space upon returning from the function, we
restore SP from BP before returning. BP thus becomes the frame
pointer, which doesn't change throughout the function call, and we
expect any displacements of BP to never reach below SP.

We also fix the calling protocol to not do the strange
environment-pointing SP manipulations: just refer to the function
arguments using positive displacements from BP (because the caller
pushed them onto the stack, and so they live on the caller's stack
frame).

The calling convention now resembles cdecl, except that:
 - callee is responsible for preserving BP;
 - we additionally push DI on the stack (throughout the call,
   DI refers to the closure object we're calling, so also needs
   to be preserved); DI is pushed before the arguments and so
   the expression being called is now also evaluated before
   the arguments – this is consistent with Clojure).

In line with cdecl, caller is responsible for cleaning up the
stack, and arguments are pushed right-to-left.

 [1]: https://lobste.rs/s/vnwuez/dos_debugging_quirk_2014