Rewrite compiler in state-passing 'monadic' style
Up until now, the contract for `compile-expr` was that it would
produce a sequence of assembly instructions _only_ for the
compiled expression; these were then combined using `concat`,
typically wrapped in `codeseq`.

With this commit:

- `compile-expr` always produces a new state, _adding_ new
  instructions into the old state's :code (and potentially
  modifying other things);
- thus, compilation works by threading initial state through
  a series of transformations, each conj'ing new instructions
  onto :code – no more concat;
- `codeseq` becomes the pervasive mechanism for state manipulation:
  everything uses it, and it now takes `state` as the initial arg;
- as a bonus, we keep track of the min :stack-pointer encountered
  within a closure (we'll make use of it later).

In the end, I quite like how this turned out. While the impl of
`codeseq` is more complicated, many other things became simpler,
and we no longer need to discriminate between state map / vector
of instructions.