Assembler: support label immediates with constant displacement
This adds support for assembly like

    :foo
    [:mov :ax [:+ :foo 17]]

which will generate code that sets AX to the address of label :foo
increased by 17. The imm argument needs to be a vector whose first
element is the symbol :+, second element is the label, and third
element is the displacement. [:+ 17 :foo] won't work.

To support this, I had to change the interim label format from a
named keyword to a map, but this should be transparent to the
compiler.