AST: support long value literals
A 'long value' is one whose content can't fit into a register.
Literals representing such values need to be therefore converted
into bytes that are part of the compiled blob.

We repurpose collect-closures to also collect long literals.
They are gathered as part of the resulting top-level `labels`
expression, and replaced with refs (expressions of type
:long-value-ref identified by a keyword).